+++ lib/i18n_alchemy/proxy.rb
@@ -2,54 +2,21 @@ module I18n
   module Alchemy
     # Depend on AS::BasicObject which has a "blank slate" - no methods.
     class Proxy < ActiveSupport::BasicObject
+      include AttributesParsing
+
       # TODO: cannot assume _id is always a foreign key.
       # Find a better way to find that and skip these columns.
       def initialize(target, attributes=nil, *args)
         @target = target
+
         @localized_attributes = {}
+        build_attributes
+        build_methods
 
-        @target.class.columns.each do |column|
-          next if column.primary || column.name.ends_with?("_id")
-
-          parser = detect_parser(column)
-          if parser
-            create_localized_attribute(column.name, parser)
-            define_localized_methods(column.name)
-          end
-        end
-
-        @localized_associations = @target.class.nested_attributes_options.map do |association_name, options|
-          ::I18n::Alchemy::AssociationParser.new(@target.class, association_name)
-        end
+        @localized_associations = []
+        build_associations
 
         assign_attributes(attributes, *args) if attributes
-      end
-
-      def attributes=(attributes)
-        @target.attributes = parse_attributes(attributes)
-      end
-
-      # This method is added to the proxy even thought it does not exist in
-      # Rails 3.0 (only >= 3.1).
-      def assign_attributes(attributes, *args)
-        if @target.respond_to?(:assign_attributes)
-          @target.assign_attributes(parse_attributes(attributes), *args)
-        else
-          self.attributes = attributes
-        end
-      end
-
-      def update_attributes(attributes, *args)
-        @target.update_attributes(parse_attributes(attributes), *args)
-      end
-
-      def update_attributes!(attributes, *args)
-        @target.update_attributes!(parse_attributes(attributes), *args)
-      end
-
-      def update_attribute(attribute, value)
-        attributes = parse_attributes(attribute => value)
-        @target.update_attribute(attribute, attributes.values.first)
       end
 
       # Override to_param to always return the +proxy.to_param+. This allow us
@@ -83,53 +50,78 @@ module I18n
 
       private
 
+      def build_attributes
+        @target.class.columns.each do |column|
+          next if column.primary || column.name.ends_with?("_id")
+
+          parser = detect_parser_from_column(column)
+          build_attribute(column.name, parser)
+        end
+      end
+
+      def build_methods
+        @target.class.localized_methods.each_pair do |method, parser_type|
+          method = method.to_s
+          parser = detect_parser(parser_type)
+          build_attribute(method, parser)
+        end
+      end
+
+      def build_associations
+        @target.class.nested_attributes_options.each_key do |association_name|
+          create_localized_association(association_name)
+        end
+      end
+
+      def build_attribute(name, parser)
+        return unless parser
+        create_localized_attribute(name, parser)
+        define_localized_methods(name)
+      end
+
+      def create_localized_association(association_name)
+        @localized_associations <<
+          AssociationParser.new(@target.class, association_name)
+      end
+
       def create_localized_attribute(column_name, parser)
         @localized_attributes[column_name] =
-          ::I18n::Alchemy::Attribute.new(@target, column_name, parser)
+          Attribute.new(@target, column_name, parser)
       end
 
       def define_localized_methods(column_name)
+        target = @target
         class << self; self; end.instance_eval do
           define_method(column_name) do
             @localized_attributes[column_name].read
           end
 
           # Before type cast must be localized to integrate with action view.
-          define_method("#{column_name}_before_type_cast") do
+          method_name = "#{column_name}_before_type_cast"
+          define_method(method_name) do
             @localized_attributes[column_name].read
-          end
+          end if target.respond_to?(method_name)
 
-          define_method("#{column_name}=") do |value|
+          method_name = "#{column_name}="
+          define_method(method_name) do |value|
             @localized_attributes[column_name].write(value)
-          end
+          end if target.respond_to?(method_name)
         end
       end
 
-      def detect_parser(column)
-        case
-        when column.number?
+      def detect_parser_from_column(column)
+        detect_parser(column.number? ? :number : column.type)
+      end
+
+      def detect_parser(type)
+        case type
+        when :number
           NumericParser
-        when column.type == :date
+        when :date
           DateParser
-        when column.type == :datetime || column.type == :timestamp
+        when :datetime, :timestamp
           TimeParser
         end
-      end
-
-      def parse_attributes(attributes)
-        attributes = attributes.stringify_keys
-
-        @localized_attributes.each do |column_name, attribute|
-          next unless attributes.key?(column_name)
-          attributes[column_name] = attribute.parse(attributes[column_name])
-        end
-        @localized_associations.each do |association_parser|
-          association_attributes = association_parser.association_name_attributes
-          next unless attributes.key?(association_attributes)
-          attributes[association_attributes] = association_parser.parse(attributes[association_attributes])
-        end
-
-        attributes
       end
     end
   end