+++ test/i18n_alchemy/numeric_parser_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-class NumericParserTest < MiniTest::Unit::TestCase
+class NumericParserTest < I18n::Alchemy::TestCase
   def setup
     @parser = I18n::Alchemy::NumericParser
   end